refactor: import 경로 변경으로 인한 수정 및 PriceSection prop 수정
